requirements.txt
cython
numpy
pandas
matplotlib
scikit-learn
pydot
pillow
tensorflow-cpu==2.1
